requirements.txt
cssselect==0.9.1
cssutils==1.0
data-importer==2.2.1
diff-match-patch==20121119
dj-database-url==0.3.0
Django==1.8.1
django-absolute==0.3
django-bootstrap3-datetimepicker==2.2.3
django-emailit==0.2.1
django-import-export==0.2.7
django-tables2==0.16.0
gunicorn==19.3.0
jdcal==1.0
lxml==3.4.4
openpyxl==2.2.2
premailer==2.8.4
psycopg2==2.5.4
six==1.9.0
tablib==0.10.0
virtualenv==12.1.1
whitenoise==1.0.6
xlrd==0.9.3
